+++ internal/driver/ansible/ansiblePlaybookDriver.go
@@ -10,7 +10,7 @@ import (
 	ansible "github.com/apenella/go-ansible/pkg/playbook"
 	errors "github.com/apenella/go-common-utils/error"
 	"github.com/gostevedore/stevedore/internal/build/varsmap"
-	"github.com/gostevedore/stevedore/internal/types"
+	"github.com/gostevedore/stevedore/internal/driver"
 )
 
 const (
@@ -48,7 +48,7 @@ func NewAnsiblePlaybookDriver(driver AnsibleDriverer, writer io.Writer) (*Ansibl
 }
 
 // Build performs the build. In case the build could not performed it returns an error
-func (d *AnsiblePlaybookDriver) Build(ctx context.Context, o *types.BuildOptions) error {
+func (d *AnsiblePlaybookDriver) Build(ctx context.Context, o *driver.BuildDriverOptions) error {
 
 	builderName := "builder"
 	errContext := "(ansibledriver::Build)"