+++ internal/driver/docker/godockerbuilder/interface.go
@@ -10,17 +10,18 @@ import (
 
 // DockerBuilder defines a docker driver
 type DockerBuilder interface {
-	WithDockerfile(dockerfile string) *build.DockerBuildCmd
-	WithImageName(image string) *build.DockerBuildCmd
+	WithDockerfile(string) *build.DockerBuildCmd
+	WithImageName(string) *build.DockerBuildCmd
+	WithPullParentImage() *build.DockerBuildCmd
 	WithPushAfterBuild() *build.DockerBuildCmd
-	WithResponse(response types.Responser) *build.DockerBuildCmd
+	WithResponse(types.Responser) *build.DockerBuildCmd
 	WithUseNormalizedNamed() *build.DockerBuildCmd
 	WithRemoveAfterPush() *build.DockerBuildCmd
-	AddAuth(username string, password string, registry string) error
-	AddPushAuth(username string, password string) error
-	AddBuildArgs(arg string, value string) error
-	AddBuildContext(contexts ...godockerbuilderbuildcontext.DockerBuildContexter) error
-	AddLabel(label string, value string) error
-	AddTags(tags ...string) error
-	Run(ctx context.Context) error
+	AddAuth(string, string, string) error
+	AddPushAuth(string, string) error
+	AddBuildArgs(string, string) error
+	AddBuildContext(...godockerbuilderbuildcontext.DockerBuildContexter) error
+	AddLabel(string, string) error
+	AddTags(...string) error
+	Run(context.Context) error
 }